refactor: restructure auth-context location
Moved `auth-context` from `contexts` to `features/auth/contexts`. Updated all imports accordingly for consistency and better modularity. This improves the organization of feature-specific code.
