+++ apps/nextjs/src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import type { PropsWithChildren } from 'react';
 import { LayoutDefault } from '@/components/ui/layout';
 import { AppContext } from '@/contexts/app-context';
-import { AuthContext } from '@/contexts/auth-context';
+import { AuthContext } from '@/features/auth/contexts/auth-context';
 import { description } from '@/lib/metatag/description';
 import { title } from '@/lib/metatag/title';
 import { roboto } from '@/styles/fonts';
