--- apps/nextjs/src/contexts/auth-context/index.ts
@@ -1,1 +1,0 @@
-export { AuthContext, useAuthContext } from './auth-context';
